+++ app/Services/ProductService.php
@@ -23,6 +23,58 @@ class ProductService
         $this->uploader = new Uploader();
     }
 
+    public function searchProducts(
+        string $search,
+        $page,
+        array $filter_cats = null,
+        $filter_price_min = null,
+        $filter_price_max = null,
+        $filter_rating = null,
+        bool $sort_newest = false,
+        bool $sort_sell = false,
+        bool $sort_desc_price = null
+    ) {
+        $per_page = 12;
+        $key_words = $search ? explode(" ", $search) : [];
+        $products = $this->product_repository->searchProducts(
+            $key_words,
+            $filter_cats,
+            $filter_price_min,
+            $filter_price_max,
+            $filter_rating,
+            $sort_newest,
+            $sort_sell,
+            $sort_desc_price,
+        );
+        $num_page = ceil($products->count() / $per_page);
+        $data["num_page"] = $num_page;
+        $data["products"] = $products->slice(($page - 1) * $per_page, $per_page);
+        return $data;
+    }
+
+    public function getRecommendedProducts($page)
+    {
+        return $this->product_repository->getRecommendedProducts($page);
+    }
+
+    public function getTopSellingProducts()
+    {
+        return $this->product_repository->getTopSellingProducts();
+    }
+
+    public function getFeaturedProducts()
+    {
+        return $this->product_repository->getFeaturedProducts();
+    }
+
+    public function getDetails($slug)
+    {
+        if ($product = $this->product_repository->findBySlug($slug)) {
+            return $product;
+        }
+        return false;
+    }
+
     public function updateRating($id)
     {
         if ($product = $this->find($id)) {
@@ -40,6 +92,10 @@ class ProductService
     public function delete($id)
     {
         $product = $this->find($id);
+        $shop = auth()->user()->shop;
+        if (!$product || !$shop || $product->shop_id != $shop->id) {
+            return false;
+        }
         if ($product) {
             foreach ($product->variants as $variants) {
                 if ($variants->colorImage) {
@@ -58,7 +114,7 @@ class ProductService
             }
             return $product->delete();
         }
-        return $product;
+        return true;
     }
 
     public function updateOrCreate(array $data, array $variant_keys, array $discount_keys)
@@ -69,6 +125,9 @@ class ProductService
         }
 
         if (Arr::exists($data, "id") && $product = $this->find($data["id"])) {
+            if ($product->shop_id != $data["shop_id"]) {
+                return false;
+            }
             $old_image_ids = $product->image_ids;
             foreach ($old_image_ids as $id) {
                 $this->uploader->delete($id);
